fix: Various btn fixes
