+++ dev/components/components/button-dropdown.vue
@@ -10,7 +10,7 @@
             <q-item-tile label>Photos</q-item-tile>
             <q-item-tile sublabel>February 22, 2016</q-item-tile>
           </q-item-main>
-          </q-item-side right icon="info" />
+          <q-item-side right icon="info" />
         </q-item>
         <q-item-separator inset />
         <q-list-header inset>Files</q-list-header>
@@ -20,7 +20,7 @@
             <q-item-tile label>Vacation</q-item-tile>
             <q-item-tile sublabel>February 22, 2016</q-item-tile>
           </q-item-main>
-          </q-item-side right icon="info" />
+          <q-item-side right icon="info" />
         </q-item>
       </q-list>
     </q-btn-dropdown>
@@ -39,7 +39,7 @@
                 <q-item-tile label>Photos</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
             <q-item-separator inset />
             <q-list-header inset>Files</q-list-header>
@@ -49,7 +49,7 @@
                 <q-item-tile label>Vacation</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
           </q-list>
         </q-btn-dropdown>
@@ -62,7 +62,7 @@
                 <q-item-tile label>Photos</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
             <q-item-separator inset />
             <q-list-header inset>Files</q-list-header>
@@ -72,7 +72,7 @@
                 <q-item-tile label>Vacation</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
           </q-list>
         </q-btn-dropdown>
@@ -85,7 +85,7 @@
                 <q-item-tile label>Photos</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
             <q-item-separator inset />
             <q-list-header inset>Files</q-list-header>
@@ -95,7 +95,7 @@
                 <q-item-tile label>Vacation</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
           </q-list>
         </q-btn-dropdown>
@@ -108,7 +108,7 @@
                 <q-item-tile label>Photos</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
             <q-item-separator inset />
             <q-list-header inset>Files</q-list-header>
@@ -118,7 +118,7 @@
                 <q-item-tile label>Vacation</q-item-tile>
                 <q-item-tile sublabel>February 22, 2016</q-item-tile>
               </q-item-main>
-              </q-item-side right icon="info" />
+              <q-item-side right icon="info" />
             </q-item>
           </q-list>
         </q-btn-dropdown>
