+++ src/components/btn/btn-mixin.js
@@ -1,11 +1,13 @@
 import Ripple from '../../directives/ripple'
 import { QIcon } from '../icon'
+import AlignMixin from '../../mixins/align'
 
 const sizes = {
   xs: 8, sm: 10, md: 14, lg: 20, xl: 24
 }
 
 export default {
+  mixins: [AlignMixin],
   components: {
     QIcon
   },
@@ -32,11 +34,6 @@ export default {
     glossy: Boolean,
     dense: Boolean,
     noRipple: Boolean,
-    justify: {
-      type: String,
-      default: 'center',
-      validator: v => ['start', 'end', 'center', 'between', 'around'].includes(v)
-    },
     tabindex: Number
   },
   computed: {
@@ -112,7 +109,7 @@ export default {
       return cls
     },
     innerClasses () {
-      const classes = [`justify-${this.justify}`]
+      const classes = [ this.alignClass ]
       if (this.noWrap) {
         classes.push('no-wrap', 'text-no-wrap')
       }
