+++ src/css/core/positioning.styl
@@ -1,6 +1,6 @@
-.pull-left, .float-left
+.float-left
   float left
-.pull-right, .float-right
+.float-right
   float right
 
 .relative-position
